+++ src/c_lib/common/files.hpp
@@ -7,7 +7,7 @@
 
 int GS_MKDIR(const char* dir, int permissions);
 
-char* read_file_to_buffer(const char* file_name, int* size);
+char* read_file_to_buffer(const char* file_name, size_t* size);
 char* fast_read_file_to_buffer(const char* file_name, int* size);
 
 off_t fsize(const char *filename);
